+++ tests/testthat/replication.R
@@ -15,7 +15,7 @@ correlation <- function(a, b) {
     )
 }
 
-corp <- data_corpus_inaugural %>% 
+corp <- head(data_corpus_inaugural, 59) %>% 
     corpus_reshape()
 
 toks <- tokens(corp, remove_punct = TRUE, remove_symbols = TRUE) %>%